requirements.txt
streamlit
web3==6.*
python-dotenv
pandas
plotly
sqlalchemy
asyncio
